+++ src/004-largest-palindrome-product/c/src/solve.h
@@ -1,7 +1,4 @@
-#include <stddef.h>
 #include <stdint.h>
-#include <math.h>
-#include "euler/palindrome.h"
 #pragma once
 
-uint32_t solve(size_t digits);
+uint32_t solve(uint32_t digits);
