+++ src/005-smallest-multiple/c/src/solve.h
@@ -1,6 +1,4 @@
-#include <stddef.h>
 #include <stdint.h>
-#include <euler/gcd_lcm.h>
 #pragma once
 
-uint32_t solve(size_t digits);
+uint32_t solve(uint32_t digits);
