+++ src/007-10001st-prime/c/src/solve.c
@@ -1,7 +1,7 @@
 #include <euler/simple_prime.h>
 #include "solve.h"
 
-uint64_t solve(size_t pos)
+uint64_t solve(uint32_t pos)
 {
     prime_t p = prime_new();
     uint64_t prime = prime_nth(&p, pos-1);
