+++ src/007-10001st-prime/c/src/solve.h
@@ -1,5 +1,4 @@
 #include <stdint.h>
-#include <stddef.h>
 #pragma once
 
-uint64_t solve(size_t pos);
+uint64_t solve(uint32_t pos);
