+++ src/011-largest-product-in-a-grid/c/src/solve.h
@@ -1,8 +1,8 @@
-#include "euler/vec.h"
 #include <stdint.h>
 #include <stddef.h>
+#include "euler/vec.h"
 #pragma once
 
-uint64_t get(vec_t *grid, int x, int y);
+uint64_t get(vec_t *grid, int64_t x, int64_t y);
 uint64_t solve(vec_t *v, size_t len);
-uint64_t max_product(vec_t *grid, int dir_x, int dir_y, size_t len);
+uint64_t max_product(vec_t *grid, int8_t dir_x, int8_t dir_y, size_t len);
